[APP-2479] Extract the new typography & make direct usage of it
