+++ app-games/src/main/java/com/aptoide/android/aptoidegames/feature_apps/presentation/CarouselLargeAppView.kt
@@ -46,7 +46,7 @@ import com.aptoide.android.aptoidegames.home.getSeeMoreRouteNavigation
 import com.aptoide.android.aptoidegames.installer.presentation.AppIconWProgress
 import com.aptoide.android.aptoidegames.installer.presentation.InstallViewShort
 import com.aptoide.android.aptoidegames.installer.presentation.ProgressText
-import com.aptoide.android.aptoidegames.theme.AppTheme
+import com.aptoide.android.aptoidegames.theme.AGTypography
 import com.aptoide.android.aptoidegames.theme.AptoideTheme
 import com.aptoide.android.aptoidegames.theme.Palette
 
@@ -182,7 +182,7 @@ private fun CarouselLargeAppView(
           color = Palette.White,
           maxLines = 2,
           overflow = TextOverflow.Ellipsis,
-          style = AppTheme.typography.descriptionGames
+          style = AGTypography.DescriptionGames
         )
         ProgressText(
           app = app,
