+++ app-games/src/main/java/com/aptoide/android/aptoidegames/settings/SettingsScreen.kt
@@ -51,7 +51,7 @@ import com.aptoide.android.aptoidegames.drawables.icons.getForward
 import com.aptoide.android.aptoidegames.network.presentation.NetworkPreferencesViewModel
 import com.aptoide.android.aptoidegames.terms_and_conditions.ppUrl
 import com.aptoide.android.aptoidegames.terms_and_conditions.tcUrl
-import com.aptoide.android.aptoidegames.theme.AppTheme
+import com.aptoide.android.aptoidegames.theme.AGTypography
 import com.aptoide.android.aptoidegames.theme.AptoideTheme
 import com.aptoide.android.aptoidegames.theme.Palette
 import com.aptoide.android.aptoidegames.toolbar.AppGamesTopBar
@@ -158,7 +158,7 @@ fun SettingsViewContent(
           )
           Text(
             text = stringResource(R.string.settings_about_version, verName, verCode),
-            style = AppTheme.typography.inputs_S,
+            style = AGTypography.InputsS,
             modifier = Modifier
               .defaultMinSize(minHeight = 48.dp)
               .wrapContentHeight()
@@ -182,7 +182,7 @@ fun SettingsViewContent(
             Text(
               text = hardwareSpecsText,
               modifier = Modifier.weight(weight = 1f),
-              style = AppTheme.typography.inputs_S,
+              style = AGTypography.InputsS,
               color = Palette.GreyLight
             )
             AppGamesOutlinedButton(
@@ -217,7 +217,7 @@ fun SettingsViewContent(
         ) {
           Text(
             text = "Powered by",
-            style = AppTheme.typography.body,
+            style = AGTypography.Body,
             color = Palette.GreyLight
           )
         }
@@ -255,14 +255,14 @@ fun SettingsSectionHeader(
 ) {
   Text(
     text = title,
-    style = AppTheme.typography.title,
+    style = AGTypography.Title,
     modifier = Modifier.padding(start = 24.dp, top = 16.dp),
     color = Palette.White
   )
   subTitle?.let {
     Text(
       text = it,
-      style = AppTheme.typography.smallGames,
+      style = AGTypography.SmallGames,
       modifier = Modifier.padding(start = 24.dp),
       color = Palette.GreyLight
     )
@@ -302,7 +302,7 @@ fun SettingsSwitchItem(
     Text(
       text = title,
       modifier = Modifier.weight(weight = 1f),
-      style = AppTheme.typography.inputs_S,
+      style = AGTypography.InputsS,
       color = Palette.GreyLight
     )
     AptoideGamesSwitch(
@@ -330,7 +330,7 @@ fun SettingsCaretItem(
       Text(
         text = title,
         modifier = Modifier.weight(weight = 1f),
-        style = AppTheme.typography.inputs_S,
+        style = AGTypography.InputsS,
         color = Palette.GreyLight
       )
     } else {
@@ -343,12 +343,12 @@ fun SettingsCaretItem(
         Text(
           text = title,
           modifier = Modifier.padding(bottom = 11.dp),
-          style = AppTheme.typography.inputs_S,
+          style = AGTypography.InputsS,
           color = Palette.White
         )
         Text(
           text = subtitle,
-          style = AppTheme.typography.inputs_S,
+          style = AGTypography.InputsS,
           color = Palette.White
         )
       }
